refactor(admin): ui 폴더 파일 네이밍 변경
