+++ src/features/admin/ui/container/index.ts
@@ -1,0 +1,1 @@
+export * from './TableContainer';
