+++ src/features/admin/ui/form/AddItemForm.tsx
@@ -5,8 +5,8 @@ import { toast } from 'sonner';
 
 import { Button, Checkbox, Form, Input } from '@/shared';
 
-import { ItemField } from '../components';
-import { AddItem, AddItemSchema, ITEM_TYPE } from '../model';
+import { ItemField } from '../../components';
+import { AddItem, AddItemSchema, ITEM_TYPE } from '../../model';
 import { zodResolver } from '@hookform/resolvers/zod';
 
 export const AddItemForm = () => {
