+++ src/features/admin/ui/form/index.ts
@@ -1,0 +1,1 @@
+export * from './AddItemForm';
