+++ src/features/admin/ui/index.ts
@@ -1,7 +1,3 @@
-export * from './table-container';
-export * from './add-item-form';
-export * from './lender-info-section';
-export * from './item-name-section';
-export * from './item-type-section';
-export * from './item-period-section';
-export * from './item-image-section';
+export * from './form';
+export * from './section';
+export * from './container';
