+++ src/features/admin/ui/section/LenderInfoSection.tsx
@@ -1,4 +1,4 @@
-import { ItemWrapper } from '../components';
+import { ItemWrapper } from '../../components';
 
 type Props = {
   name: string;
